SRS general template
Check the TOC of the document. It includes brief descriptions what should be added in each point. Until we determine which points will be required, we can start filling 1 and 2 sections. I added clarifications in most points.
I guess point 3 will be related to the test plan we will create. We can remove points.

**tracking changes in a DOCX file in git is an awful thing**